+++ webots_ros2_driver/include/webots_ros2_driver/plugins/static/Ros2LED.hpp
@@ -15,18 +15,16 @@
 #ifndef ROS2_LED_HPP
 #define ROS2_LED_HPP
 
-#include <unordered_map>
-#include <std_msgs/msg/int32.hpp>
 #include <webots/led.h>
-#include <webots_ros2_driver/utils/Utils.hpp>
-#include <webots_ros2_driver/WebotsNode.hpp>
+#include <std_msgs/msg/int32.hpp>
+#include <unordered_map>
 #include <webots_ros2_driver/PluginInterface.hpp>
+#include <webots_ros2_driver/WebotsNode.hpp>
+#include <webots_ros2_driver/utils/Utils.hpp>
 
-namespace webots_ros2_driver
-{
+namespace webots_ros2_driver {
 
-  class Ros2LED : public PluginInterface
-  {
+  class Ros2LED : public PluginInterface {
   public:
     void init(webots_ros2_driver::WebotsNode *node, std::unordered_map<std::string, std::string> &parameters) override;
     void step() override;
@@ -39,6 +37,6 @@ namespace webots_ros2_driver
     rclcpp::Subscription<std_msgs::msg::Int32>::SharedPtr mSubscriber;
   };
 
-}
+}  // namespace webots_ros2_driver
 
 #endif